+++ straxen/storage/README.md
@@ -1,0 +1,6 @@
+# Straxen storage
+In this subfolder there are two types of storages:
+ - [`strax.StorageFrontends`](https://strax.readthedocs.io/en/latest/developer/storage.html) -> XENONnT specific code for managing straxens data
+ - [`straxen.Downloader`](https://straxen.readthedocs.io/en/latest/config_storage.html) -> This downloads configuration files stored in the mongo database
+
+The associated tests are written under `straxen/tests/storage`.
